Use sqlalchemy 1.4
Use asyncpg instead of aiopg.
Note that this required specifying dates as datetime.datetime
instead of ISO strings, which required some extra work to compare
messages added to the database with messages returned by the service.

Also fix a bug in find_messages: the sort order was not deterministic
which could lead to returning the wrong messages when using
offset and limit. Fixed by appending "id" to order_by unless it
already appears (as "id" or "-id")

Greatly expanded the unit test coverage.