Upgrade python, Butler, fastapi, and pydantic
Upgrade dependencies to get a new Butler version with the 'LabeledButlerFactory' class available, needed to fix threadsafety issues in DM-43119.  

Butler now requires pydantic v2, so we also need to update FastAPI to a version that supports v2.  

Butler only supports Python >=3.11, so bump the Python version as well.  This in turn required upgrading PyYaml for compatibility.