Fix compatibility with Butler universe 6
The Butler repositories used by exposurelog will be updated to a new database schema soon.  The new schema removes the the "group_id" property from exposure, and renames "group_name" to "group.

This affects the Exposure objects returned by the /exposures/ endpoint.  It is believed that group_id is unused by downstream consumers of exposurelog, so it has been removed entirely.  group_name is used by LOVE, so we now pull this property from whichever of "group_name" or "group" is available.